Disable Javadoc lint warnings in Gradle build (#1128)
Most of our code is auto-generated, which typically includes Javadoc
comments. Therefore, the Javadoc lint warnings that are triggered during
the build process are not meaningful. This commit adds a configuration
to the build.gradle.kts file to disable these warnings, reducing noise
in the build output and making it easier to spot other potential issues.