Allow to git tag containing `v` as prefix (#1481)
We've decided to specify git tag versions with a `v` prefix for all
bot-sdk repositories. This is to apply the same rule across all
repositories and reduce mistakes.

In the case of line-bot-sdk-java, specifying the `v` prefix is not an
issue as long as it is removed during publishing. This change supports
it.